refactor: switch client↔server wire to consensus encoding
Mirror the p2p layer by sending IrohApiRequest and its response bytes
through Encodable/Decodable instead of JSON. Endpoint Param/Response now
require Encodable/Decodable (drop DeserializeOwned/Serialize). Params
carry consensus-encoded bytes in ApiRequestErased; Result<Vec<u8>,
ApiError> is the wire response shape.

Delete SerdeModuleEncoding*, SerdeTransaction, ApiRequest<T> — transactions
and module responses travel as their typed values directly. usize params
(await_incoming_contracts) widen to u64 since usize isn't Encodable,
OutputInfo gains Encodable/Decodable derives, ApiError.code widens to u32.

+++ modules/fedimint-lnv2-client/src/api.rs
@@ -28,7 +28,7 @@ pub trait LightningFederationApi {
 
     async fn await_preimage(&self, contract_id: OutPoint, expiration: u64) -> Option<[u8; 32]>;
 
-    async fn await_incoming_contracts(&self, start: u64, n: usize) -> (Vec<IncomingContract>, u64);
+    async fn await_incoming_contracts(&self, start: u64, n: u64) -> (Vec<IncomingContract>, u64);
 
     async fn gateways(&self) -> FederationResult<Vec<SafeUrl>>;
 
@@ -68,7 +68,7 @@ where
         .await
     }
 
-    async fn await_incoming_contracts(&self, start: u64, n: usize) -> (Vec<IncomingContract>, u64) {
+    async fn await_incoming_contracts(&self, start: u64, n: u64) -> (Vec<IncomingContract>, u64) {
         self.request_current_consensus_retry(
             AWAIT_INCOMING_CONTRACTS_ENDPOINT.to_string(),
             ApiRequestErased::new((start, n)),
